+++ src-docs/src/views/color_picker/color_stops.js
@@ -1,6 +1,14 @@
 import React, { useState } from 'react';
 
-import { EuiColorStops, EuiFormRow } from '../../../../src/components';
+import {
+  EuiColorStops,
+  EuiFormRow,
+  EuiRange,
+  EuiFlexGroup,
+  EuiFlexItem,
+  EuiButtonEmpty,
+  EuiPopover,
+} from '../../../../src/components';
 
 import { useColorStopsState } from '../../../../src/services';
 
@@ -12,6 +20,9 @@ export default () => {
     addRandomColor,
   ] = useColorStopsState(true);
   const [fixedColorStops, setFixedColorStops] = useColorStopsState(true);
+  const [steppedColorStops, setSteppedColorStops] = useColorStopsState(true);
+  const [value, setValue] = useState(10);
+  const [isPopoverOpen, setIsPopoverOpen] = useState(false);
 
   const [extendedColorStops, setExtendedColorStops] = useState([
     {
@@ -37,6 +48,21 @@ export default () => {
   const handleEmptyChange = (colorStops) => {
     setEmptyColorStops(colorStops);
   };
+
+  const onButtonClick = () =>
+    setIsPopoverOpen((isPopoverOpen) => !isPopoverOpen);
+  const closePopover = () => setIsPopoverOpen(false);
+
+  const button = (
+    <EuiButtonEmpty
+      onClick={onButtonClick}
+      iconType="controlsVertical"
+      aria-label="Open settings"
+      color="text"
+      size="xs">
+      Steps
+    </EuiButtonEmpty>
+  );
 
   return (
     <React.Fragment>
@@ -86,6 +112,41 @@ export default () => {
           max={100}
           stopType="fixed"
         />
+      </EuiFormRow>
+
+      <EuiFormRow label="Stepped color segments">
+        <EuiFlexGroup alignItems="center" gutterSize="xs">
+          <EuiFlexItem>
+            <EuiColorStops
+              label="Stepped color segments"
+              onChange={setSteppedColorStops}
+              colorStops={steppedColorStops}
+              stepNumber={value}
+              min={0}
+              max={100}
+              stopType="stepped"
+            />
+          </EuiFlexItem>
+
+          <EuiFlexItem grow={false}>
+            <EuiPopover
+              panelStyle={{ minWidth: 380 }}
+              button={button}
+              isOpen={isPopoverOpen}
+              closePopover={closePopover}>
+              <EuiFormRow label="Number of steps" display="columnCompressed">
+                <EuiRange
+                  value={value}
+                  onChange={(e) => setValue(parseInt(e.target.value))}
+                  showInput
+                  aria-label="Change the number of steps"
+                  min={2}
+                  max={20}
+                />
+              </EuiFormRow>
+            </EuiPopover>
+          </EuiFlexItem>
+        </EuiFlexGroup>
       </EuiFormRow>
     </React.Fragment>
   );